requirements.txt
pytorch==1.0.0
numpy
progressbar2
cython
tensorboardx